+++ Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flask = "*"
+requests = "*"
+faker = "*"
+webargs = "*"
+psutil = "*"
+memory-profiler = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.10"
