+++ hw_3/templates/s3_csv_table.html
@@ -1,0 +1,9 @@
+<table id="s3_csv_table">
+{% for row in csv %}
+  <tr>
+  {% for col in row.split(";") %}
+  <td>{{ col }}</td>
+  {% endfor %}
+  </tr>
+{% endfor %}
+</table>
