+++ hw_3/templates/table_from_list.html
@@ -1,0 +1,9 @@
+<table id="table_from_list">
+{% for row in csv %}
+  <tr>
+  {% for col in row %}
+  <td>{{ col }}</td>
+  {% endfor %}
+  </tr>
+{% endfor %}
+</table>
